- refactored to int64 and int 32 - refactored to emit pointers for null types, so no more pgtype.Int8 just *int64

+++ internal/DTO/requests.go
@@ -8,15 +8,15 @@ type BrandRequest struct {
 type CategoryRequest struct {
 	Name     string `json:"name"`
 	Slug     string `json:"slug"`
-	ParentId *int   `json:"parent_id,omitempty"`
+	ParentId *int64 `json:"parent_id,omitempty"`
 }
 
 type ProductRequest struct {
-	BrandId     int     `json:"brand_id"`
-	CategoryId  int     `json:"category_id"`
+	BrandId     int64   `json:"brand_id"`
+	CategoryId  int64   `json:"category_id"`
 	Name        string  `json:"name"`
 	Slug        string  `json:"slug"`
 	Description *string `json:"description,omitempty"`
-	PriceKopeck int     `json:"price_kopeck"`
+	PriceKopeck int32   `json:"price_kopeck"`
 	IsActive    *bool   `json:"is_active,omitempty"`
 }
