+++ internal/DTO/responses.go
@@ -3,7 +3,7 @@ package DTO
 import "time"
 
 type BrandResponse struct {
-	Id        int       `json:"id"`
+	Id        int64     `json:"id"`
 	Name      string    `json:"name"`
 	Slug      string    `json:"slug"`
 	CreatedAt time.Time `json:"created_at"`
@@ -11,22 +11,22 @@ type BrandResponse struct {
 }
 
 type CategoryResponse struct {
-	Id        int       `json:"id"`
+	Id        int64     `json:"id"`
 	Name      string    `json:"name"`
 	Slug      string    `json:"slug"`
-	ParentId  *int      `json:"parent_id"`
+	ParentId  *int64    `json:"parent_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
 
 type ProductResponse struct {
-	Id          int       `json:"id"`
-	BrandId     int       `json:"brand_id"`
-	CategoryId  int       `json:"category_id"`
+	Id          int64     `json:"id"`
+	BrandId     int64     `json:"brand_id"`
+	CategoryId  int64     `json:"category_id"`
 	Name        string    `json:"name"`
 	Slug        string    `json:"slug"`
 	Description string    `json:"description"`
-	PriceKopeck int       `json:"price_kopeck"`
+	PriceKopeck int32     `json:"price_kopeck"`
 	IsActive    bool      `json:"is_active"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
